requirements.txt
openai-whisper
sounddevice
scipy
numpy
transformers
torch
datasets
scikit-learn
accelerate
sentence-transformers
faiss-cpu
chromadb
redis
groq
pyttsx3
python-dotenv
pydantic
loguru
tabulate
setuptools